making visitor CRUD + calculation API + finishing

+++ database/migrations/2020_04_08_081145_create_collection_table.php
@@ -16,7 +16,7 @@ class CreateCollectionTable extends Migration
         Schema::create('collection', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title', 100);
-            $table->decimal('rate');
+            $table->integer('rate');
             $table->string('category', 50);
             $table->integer('quantity');
             $table->timestamps();
